planner: split casetest for pushdown
