+++ planner/core/casetest/BUILD.bazel
@@ -15,7 +15,6 @@ go_test(
     ],
     data = glob(["testdata/**"]),
     flaky = True,
-    race = "on",
     shard_count = 50,
     deps = [
         "//domain",
@@ -24,7 +23,6 @@ go_test(
         "//planner/core",
         "//planner/property",
         "//session",
-        "//sessionctx/stmtctx",
         "//sessionctx/variable",
         "//statistics/handle",
         "//testkit",
