+++ planner/core/casetest/pushdown/testdata/integration_suite_in.json
@@ -1,0 +1,89 @@
+[
+  {
+    "name": "TestPushLimitDownIndexLookUpReader",
+    "cases": [
+      // Limit should be pushed down into IndexLookUpReader, row count of IndexLookUpReader and TableScan should be 1.00.
+      "explain format = 'brief' select * from tbl use index(idx_b_c) where b > 1 limit 2,1",
+      // Projection atop IndexLookUpReader, Limit should be pushed down into IndexLookUpReader, and Projection should have row count 1.00 as well.
+      "explain format = 'brief' select * from tbl use index(idx_b_c) where b > 1 order by b desc limit 2,1",
+      // Limit should be pushed down into IndexLookUpReader when Selection on top of IndexScan.
+      "explain format = 'brief' select * from tbl use index(idx_b_c) where b > 1 and c > 1 limit 2,1",
+      // Limit should NOT be pushed down into IndexLookUpReader when Selection on top of TableScan.
+      "explain format = 'brief' select * from tbl use index(idx_b_c) where b > 1 and a > 1 limit 2,1"
+    ]
+  },
+  {
+    "name": "TestPushDownToTiFlashWithKeepOrder",
+    "cases": [
+      "explain format = 'brief' select max(a) from t",
+      "explain format = 'brief' select min(a) from t"
+    ]
+  },
+  {
+    "name": "TestPushDownToTiFlashWithKeepOrderInFastMode",
+    "cases": [
+      "explain format = 'brief' select max(a) from t",
+      "explain format = 'brief' select min(a) from t"
+    ]
+  },
+  {
+    "name": "TestPushDownProjectionForTiFlashCoprocessor",
+    "cases": [
+      "desc format = 'brief' select i * 2 from t",
+      "desc format = 'brief' select DATE_FORMAT(t, '%Y-%m-%d %H') as date from t",
+      "desc format = 'brief' select md5(s) from t; -- we do generate mpp plan, while the cost-cmp failed",
+      "desc format = 'brief' select c from t where a+1=3",
+      "desc format = 'brief' select /*+ hash_agg()*/ count(b) from  (select id + 1 as b from t)A",
+      "desc format = 'brief' select /*+ hash_agg()*/ count(*) from  (select id + 1 as b from t)A",
+      "desc format = 'brief' select /*+ hash_agg()*/ sum(b) from  (select id + 1 as b from t)A",
+      "desc format = 'brief' select /*+ stream_agg()*/ count(b) from  (select id + 1 as b from t)A",
+      "desc format = 'brief' select /*+ stream_agg()*/ count(*) from  (select id + 1 as b from t)A",
+      "desc format = 'brief' select /*+ stream_agg()*/ sum(b) from  (select id + 1 as b from t)A",
+      "desc format = 'brief' select * from (select id-2 as b from t) B join (select id-2 as b from t) A on A.b=B.b",
+      "desc format = 'brief' select * from t join (select id-2 as b from t) A on A.b=t.id",
+      "desc format = 'brief' select * from t left join (select id-2 as b from t) A on A.b=t.id",
+      "desc format = 'brief' select * from t right join (select id-2 as b from t) A on A.b=t.id",
+      "desc format = 'brief' select A.b, B.b from (select id-2 as b from t) B join (select id-2 as b from t) A on A.b=B.b",
+      "desc format = 'brief' select A.id from t as A where exists (select 1 from t where t.id=A.id)",
+      "desc format = 'brief' select A.id from t as A where not exists  (select 1 from t where t.id=A.id)",
+      "desc format = 'brief' SELECT FROM_UNIXTIME(name,'%Y-%m-%d')  FROM t;"
+    ]
+  },
+  {
+    "name": "TestPushDownProjectionForTiFlash",
+    "cases": [
+      "desc format = 'brief' select /*+ hash_agg()*/ count(b) from  (select /*+ read_from_storage(tiflash[t]) */ id + 1 as b from t)A",
+      "desc format = 'brief' select /*+ hash_agg()*/ count(*) from  (select /*+ read_from_storage(tiflash[t]) */ id + 1 as b from t)A",
+      "desc format = 'brief' select /*+ hash_agg()*/ sum(b) from  (select /*+ read_from_storage(tiflash[t]) */ id + 1 as b from t)A",
+      "desc format = 'brief' select /*+ stream_agg()*/ count(b) from  (select /*+ read_from_storage(tiflash[t]) */ id + 1 as b from t)A",
+      "desc format = 'brief' select /*+ stream_agg()*/ count(*) from  (select /*+ read_from_storage(tiflash[t]) */ id + 1 as b from t)A",
+      "desc format = 'brief' select /*+ stream_agg()*/ sum(b) from  (select /*+ read_from_storage(tiflash[t]) */ id + 1 as b from t)A",
+      "desc format = 'brief' select * from (select id-2 as b from t) B join (select id-2 as b from t) A on A.b=B.b",
+      "desc format = 'brief' select * from t join (select id-2 as b from t) A on A.b=t.id",
+      "desc format = 'brief' select * from t left join (select id-2 as b from t) A on A.b=t.id",
+      "desc format = 'brief' select * from t right join (select id-2 as b from t) A on A.b=t.id",
+      "desc format = 'brief' select A.b, B.b from (select id-2 as b from t) B join (select id-2 as b from t) A on A.b=B.b",
+      "desc format = 'brief' select A.id from t as A where exists (select 1 from t where t.id=A.id)",
+      "desc format = 'brief' select A.id from t as A where not exists  (select 1 from t where t.id=A.id)",
+      "desc format = 'brief' SELECT FROM_UNIXTIME(name,'%Y-%m-%d')  FROM t;"
+    ]
+  },
+    {
+    "name": "TestSelPushDownTiFlash",
+    "cases": [
+      "explain format = 'brief' select * from t where t.a > 1 and t.b = \"flash\" or t.a + 3 * t.a = 5",
+      "explain format = 'brief' select * from t where cast(t.a as double) + 3 = 5.1",
+      "explain format = 'brief' select * from t where b > 'a' order by convert(b, unsigned) limit 2",
+      "explain format = 'brief' select * from t where b > 'a' order by b limit 2"
+    ]
+  },
+  {
+    "name": "TestJoinNotSupportedByTiFlash",
+    "cases": [
+      "explain format = 'brief' select * from table_1 a, table_1 b where a.bit_col = b.bit_col",
+      "explain format = 'brief' select * from table_1 a left join table_1 b on a.id = b.id and dayofmonth(a.datetime_col) > 100",
+      "explain format = 'brief' select * from table_1 a right join table_1 b on a.id = b.id and dayofmonth(b.datetime_col) > 100",
+      "explain format = 'brief' select * from table_1 a join table_1 b on a.id = b.id and dayofmonth(a.datetime_col) > dayofmonth(b.datetime_col)"
+    ]
+  }
+]
